+++ README.md
@@ -3,7 +3,7 @@
 
 ## About
 
-reX is a public transport routing software based on time, locations, interests,
+reX9 is a public transport routing software based on time, locations, interests,
 and heuristics.
 
 
@@ -16,8 +16,8 @@ and further stops on the trip, and computes travel plan drafts using the [HAFAS]
 ## Setup
 
 ```shell
-git clone https://github.com/panodata/rex
-cd rex
+git clone https://github.com/panodata/rex9
+cd rex9
 python3 -m venv .venv
 source .venv/bin/activate
 pip install --editable=.[develop,docs,test]
@@ -27,7 +27,7 @@ pip install --editable=.[develop,docs,test]
 ## Usage
 
 ```shell
-rex travel --from=Oranienburg --to=Kopenhagen --stops=Fürstenberg,Mildenberg --on=do-di
+rex9 travel --from=Oranienburg --to=Kopenhagen --stops=Fürstenberg,Mildenberg --on=do-di
 ```
