+++ rex-travelplan-first.html
@@ -1,0 +1,134 @@
+
+<html>
+<head>
+<meta charset="UTF-8">
+</head>
+<body>
+<h1>Travel plan</h1>
+
+<h2>Connections from Oranienburg to Fürstenberg(Havel)</h2>
+
+<h3>Duration 0:31:00 on 2023-08-24</h3>
+
+<ul>
+<li>09:16-09:47, train RE 4354; Platforms 24/1 from Oranienburg to Fürstenberg(Havel)</li>
+</ul>
+
+<h3>Duration 0:31:00 on 2023-08-24</h3>
+
+<ul>
+<li>10:16-10:47, train RE 3508; Platforms 24/1 from Oranienburg to Fürstenberg(Havel)</li>
+</ul>
+
+<h3>Duration 0:31:00 on 2023-08-24</h3>
+
+<ul>
+<li>11:16-11:47, train RE 4356; Platforms 24/1 from Oranienburg to Fürstenberg(Havel)</li>
+</ul>
+
+<h2>Connections from Fürstenberg(Havel) to Mildenberg Friedhof, Zehdenick</h2>
+
+<h3>Duration 0:59:00 on 2023-08-25</h3>
+
+<ul>
+<li>09:12-10:11, train Bus 838; from Bahnhof, Fürstenberg (Havel) to Mildenberg Friedhof, Zehdenick</li>
+</ul>
+
+<h3>Duration 1:43:00 on 2023-08-25</h3>
+
+<ul>
+<li>10:12-10:25, train RE 4357; Platforms 1/1 from Fürstenberg(Havel) to Gransee</li>
+<li>10:25-10:28, walking; from Gransee to Bahnhof, Gransee</li>
+<li>10:41-11:02, train Bus 833; from Bahnhof, Gransee to Karlshof, Zehdenick</li>
+<li>11:02-11:07, walking; from Karlshof, Zehdenick to Mildenberger Str., Zehdenick</li>
+<li>11:53-11:55, train Bus 838; from Mildenberger Str., Zehdenick to Mildenberg Friedhof, Zehdenick</li>
+</ul>
+
+<h3>Duration 0:59:00 on 2023-08-25</h3>
+
+<ul>
+<li>11:05-12:04, train Bus 838; from Bahnhof, Fürstenberg (Havel) to Mildenberg Friedhof, Zehdenick</li>
+</ul>
+
+<h2>Connections from Mildenberg Friedhof, Zehdenick to KOPENHAGEN</h2>
+
+<h3>Duration 9:23:00 on 2023-08-26</h3>
+
+<ul>
+<li>10:43-10:46, train Bus 838; from Mildenberg Friedhof, Zehdenick to Bahnhof, Zehdenick</li>
+<li>10:46-10:49, walking; from Bahnhof, Zehdenick to Zehdenick(Mark)</li>
+<li>10:57-11:28, train RB 61114; Platforms 2/28 from Zehdenick(Mark) to Oranienburg</li>
+<li>11:46-12:23, train IC 2177; Platforms 28/2 from Oranienburg to Berlin Hbf (tief)</li>
+<li>12:38-14:22, train ICE 802; Platforms 8/5 from Berlin Hbf (tief) to Hamburg Hbf</li>
+<li>14:53-18:14, train IC 384; Platforms 6D-E/10 from Hamburg Hbf to Fredericia st</li>
+<li>18:26-20:06, train ICL50058; Platforms 9/4 from Fredericia st to Koebenhavn H</li>
+</ul>
+
+<h3>Duration 8:51:00 on 2023-08-26</h3>
+
+<ul>
+<li>12:43-12:46, train Bus 838; from Mildenberg Friedhof, Zehdenick to Bahnhof, Zehdenick</li>
+<li>12:46-12:49, walking; from Bahnhof, Zehdenick to Zehdenick(Mark)</li>
+<li>12:57-13:28, train RB 61118; Platforms 2/28 from Zehdenick(Mark) to Oranienburg</li>
+<li>13:55-14:29, train IC 2179; Platforms 28/1 from Oranienburg to Berlin Hbf (tief)</li>
+<li>14:38-16:21, train ICE 800; Platforms 8/5 from Berlin Hbf (tief) to Hamburg Hbf</li>
+<li>16:54-21:34, train IC 392; Platforms 8/2 from Hamburg Hbf to Koebenhavn H</li>
+</ul>
+
+<h3>Duration 10:12:00 on 2023-08-26</h3>
+
+<ul>
+<li>14:43-14:46, train Bus 838; from Mildenberg Friedhof, Zehdenick to Bahnhof, Zehdenick</li>
+<li>14:46-14:49, walking; from Bahnhof, Zehdenick to Zehdenick(Mark)</li>
+<li>14:57-15:28, train RB 61122; Platforms 2/28 from Zehdenick(Mark) to Oranienburg</li>
+<li>15:54-16:26, train IC 2271; Platforms 28/2 from Oranienburg to Berlin Hbf (tief)</li>
+<li>16:38-18:22, train ICE 708; Platforms 8/5 from Berlin Hbf (tief) to Hamburg Hbf</li>
+<li>18:43-20:40, train RE 11178; Platforms 11D-F/2 from Hamburg Hbf to Flensburg</li>
+<li>20:48-22:14, train IC 5764; Platforms 1/4 from Flensburg to Fredericia st</li>
+<li>22:53-00:55, train IC 70478; Platforms 10/4 from Fredericia st to Koebenhavn H</li>
+</ul>
+
+<h2>Connections from KOPENHAGEN to Oranienburg</h2>
+
+<h3>Duration 13:07:00 on 2023-08-29</h3>
+
+<ul>
+<li>18:56-20:38, train ICL 71; Platforms 5/6 from Koebenhavn H to Fredericia st</li>
+<li>20:45-22:07, train IC 5767; Platforms 7/1 from Fredericia st to Flensburg</li>
+<li>22:54-00:22, train RE 11193; Platforms 2/5 from Flensburg to Neumünster</li>
+<li>03:34-04:26, train RE 11201; Platforms 5/13 from Neumünster to Hamburg Hbf</li>
+<li>05:27-07:22, train ICE 701; Platforms 5/1 from Hamburg Hbf to Berlin Hbf (tief)</li>
+<li>07:36-08:03, train IC 2274; Platforms 8/20 from Berlin Hbf (tief) to Oranienburg</li>
+</ul>
+
+<h3>Duration 9:47:00 on 2023-08-29</h3>
+
+<ul>
+<li>23:07-05:31, train D 301; from Oerestad st to Hamburg Hbf</li>
+<li>05:52-07:42, train ICE 1003; Platforms 12/5 from Hamburg Hbf to Berlin-Spandau</li>
+<li>07:42-07:47, walking; from Berlin-Spandau to Rathaus Spandau [Bus Seegef. Str.] (S+U), Berlin</li>
+<li>07:51-08:19, train Bus X36; from Rathaus Spandau [Bus Seegef. Str.] (S+U), Berlin to Hennigsdorf Bahnhof (S), Hennigsdorf</li>
+<li>08:19-08:22, walking; from Hennigsdorf Bahnhof (S), Hennigsdorf to Hennigsdorf(b Berlin)</li>
+<li>08:35-08:54, train RB 18456; Platforms 7/24 from Hennigsdorf(b Berlin) to Oranienburg</li>
+</ul>
+
+<h3>Duration 9:24:00 on 2023-08-29</h3>
+
+<ul>
+<li>23:51-06:26, train IC 399; Platforms 7/11D-F from Koebenhavn H to Hamburg Hbf</li>
+<li>06:34-08:20, train ICE 505; Platforms 8A-F/1 from Hamburg Hbf to Berlin Hbf (tief)</li>
+<li>08:47-09:15, train RE 4354; Platforms 6/24 from Berlin Hbf (tief) to Oranienburg</li>
+</ul>
+
+<h3>Duration 9:22:00 on 2023-08-29</h3>
+
+<ul>
+<li>23:53-00:25, train R 1291; Platforms 4/3 from Koebenhavn H to Ringsted st</li>
+<li>00:41-06:26, train IC 399; Platforms 1/11D-F from Ringsted st to Hamburg Hbf</li>
+<li>06:34-08:20, train ICE 505; Platforms 8A-F/1 from Hamburg Hbf to Berlin Hbf (tief)</li>
+<li>08:47-09:15, train RE 4354; Platforms 6/24 from Berlin Hbf (tief) to Oranienburg</li>
+</ul>
+
+</body>
+</html>
+
