+++ rex9/util/common.py
